+++ tests/test_dcos_e2e/backends/docker/test_distributions.py
@@ -31,6 +31,7 @@ def _get_node_distribution(node: Node) -> Distribution:
     distributions = {
         ('"centos"', '"7"'): Distribution.CENTOS_7,
         ('coreos', '1298.7.0'): Distribution.COREOS,
+        ('ubuntu', '"16.04"'): Distribution.UBUNTU_16_04,
     }
 
     distro_id = version_data['ID'].strip()
@@ -170,6 +171,38 @@ class TestCoreOS:
     ) -> None:
         """
         DC/OS Enterprise can start up on CoreOS.
+        """
+        _enterprise_distribution_test(
+            distribution=Distribution.COREOS,
+            enterprise_artifact=enterprise_artifact,
+            license_key_contents=license_key_contents,
+        )
+
+
+class TestUbuntu1604:
+    """
+    Tests for the Ubuntu 16.04 distribution option.
+    """
+
+    def test_oss(
+        self,
+        oss_artifact: Path,
+    ) -> None:
+        """
+        DC/OS OSS can start up on Ubuntu 16.04.
+        """
+        _oss_distribution_test(
+            distribution=Distribution.COREOS,
+            oss_artifact=oss_artifact,
+        )
+
+    def test_enterprise(
+        self,
+        enterprise_artifact: Path,
+        license_key_contents: str,
+    ) -> None:
+        """
+        DC/OS Enterprise can start up on Ubuntu 16.04.
         """
         _enterprise_distribution_test(
             distribution=Distribution.COREOS,